vision_raspberry_pi/requirements.txt
numpy
opencv-python
pyserial
imutils
gpiozero

